[cleanup] Move tests/utils.py to utils.py
The context manager is going to be used outside tests as well.
